+++ DirectProgramming/C++SYCL_FPGA/ReferenceDesigns/convolution2d/src/main.cpp
@@ -15,8 +15,8 @@
 #include <sycl/sycl.hpp>
 
 #include "bmp_tools.hpp"
-#include "convolution_kernel.hpp"
 #include "exception_handler.hpp"
+#include "image_buffer_adapters.hpp"
 #include "vvp_stream_adapters.hpp"
 
 #ifndef DEFAULT_EXTENSION
@@ -39,6 +39,12 @@
 #define TEST_CONV2D_ISOLATED 0
 #endif
 
+#if TEST_CONV2D_ISOLATED
+#include "run_convolution_kernel.hpp"
+#else
+#include "run_kernel_system.hpp"
+#endif
+
 #define M_DEFAULT_INPUT DEFAULT_INPUT
 #define M_DEFAULT_OUTPUT DEFAULT_OUTPUT
 #define M_DEFAULT_EXPECTED DEFAULT_EXPECTED
@@ -48,123 +54,8 @@
 /////////////////////
 // Test subroutines
 /////////////////////
-
-/// @brief Initialize a buffer meant to store an image. This helps with
-/// debugging, because you can see if an image was only partly written. The
-/// buffer is initialized with an incrementing pattern, ranging from `0` to
-/// `(1 << 24)`.
-/// @param[out] buf Buffer to initialize
-/// @param[in] size Number of pixels to initialize in the buffer
-void InitializeBuffer(conv2d::PixelRGB *buf, size_t size) {
-  uint16_t pixel = 0;
-  for (size_t i = 0; i < size; i++) {
-    pixel++;
-    buf[i] = conv2d::PixelRGB{pixel, pixel, pixel};
-  }
-}
-
-/// @brief Convert pixels read from a bmp image using the bmp_tools functions to
-/// pixels that can be parsed by our 2D convolution IP.
-/// @param[in] bmp_img pixels read by bmptools
-/// @param[out] vvp_buf pixels to be consumed by 2D convolution IP
-/// @param[in] pixel_count (input) number of pixels in input image and output
-/// image
-void ConvertToVvpRgb(bmp_tools::BitmapRGB bmp_img, conv2d::PixelRGB *vvp_buf,
-                     size_t pixel_count) {
-  std::cout << "INFO: convert to vvp type." << std::endl;
-  for (size_t idx = 0; idx < pixel_count; idx++) {
-    uint32_t pixel_int = bmp_img(idx);
-    bmp_tools::PixelRGB bmp_rgb(pixel_int);
-
-    // convert from 8-bit to whatever the VVP IP expects
-    conv2d::PixelRGB pixel_vvp{
-        (uint16_t)(bmp_rgb.b << (conv2d::kBitsPerChannel - 8)),   //
-        (uint16_t)(bmp_rgb.g << (conv2d::kBitsPerChannel - 8)),   //
-        (uint16_t)(bmp_rgb.r << (conv2d::kBitsPerChannel - 8))};  //
-
-    vvp_buf[idx] = pixel_vvp;
-  }
-}
-
-/// @brief Convert pixels read from the 2D convolution IP to a format that can
-/// be read by the bmptools functions.
-/// @param[in] vvp_buf Pixels produced by 2D convolution IP
-/// @param[in] rows Number of rows in input image and output image
-/// @param[in] cols Number of columns in input image and output image
-/// @return bmp_buf Pixels to send to bmptools
-bmp_tools::BitmapRGB ConvertToBmpRgb(conv2d::PixelRGB *vvp_buf, size_t rows,
-                                     size_t cols) {
-  std::cout << "INFO: convert to bmp type." << std::endl;
-  size_t pixel_count = rows * cols;
-  bmp_tools::BitmapRGB bmp_img(rows, cols);
-  for (size_t idx = 0; idx < pixel_count; idx++) {
-    conv2d::PixelRGB pixel_conv = vvp_buf[idx];
-
-    // convert the VVP IP back to 8-bit
-    bmp_tools::PixelRGB bmp_rgb(
-        (uint8_t)(pixel_conv.r >> (conv2d::kBitsPerChannel - 8)),   //
-        (uint8_t)(pixel_conv.g >> (conv2d::kBitsPerChannel - 8)),   //
-        (uint8_t)(pixel_conv.b >> (conv2d::kBitsPerChannel - 8)));  //
-
-    uint32_t pixel_int = bmp_rgb.GetImgPixel();
-    bmp_img(idx) = pixel_int;
-  }
-
-  return bmp_img;
-}
-
-/// @brief Verify image dimensions from a just-read image and compare with
-/// previous image dimensions if appropriate.
-/// @param rows previous image rows (0 if no previous dimensions)
-/// @param cols previous image cols (0 if no previous dimensions)
-/// @param rows_new new image rows
-/// @param cols_new new image columns
-/// @return `true` if new dimensions are acceptable, and `rows` and `cols` have
-/// been successfully updated to `rows_new` and `cols_new` respectively.
-bool UpdateAndCheckImageDimensions(size_t &rows, size_t &cols, size_t rows_new,
-                                   size_t cols_new) {
-  // sanity check: all images should be the same size
-  if (rows == 0)
-    rows = rows_new;
-  else if (rows != rows_new) {
-    std::cerr << "ERROR: dimensions of images must match. Expected " << rows
-              << " but saw " << rows_new << "!" << std::endl;
-    return false;
-  }
-
-  if (cols == 0)
-    cols = cols_new;
-  else if (cols != cols_new) {
-    std::cerr << "ERROR: dimensions of images must match. Expected " << cols
-              << " but saw " << cols_new << "!" << std::endl;
-    return false;
-  }
-
-  // Max allowable value for rows * cols must be less than the max value of a
-  // signed 32-bit integer.
-  constexpr int kRowsColsMax = 1 << 29;
-
-  bool image_size_ok =
-      (rows_new > 0) && (cols_new > 0) && (rows_new * cols_new < kRowsColsMax);
-
-  // sanity check; this design assumes that the number of columns in the input
-  // image is a multiple of kParallelPixels.
-  if (cols % conv2d::kParallelPixels != 0) {
-    std::cerr << "ERROR: image cols = " << cols
-              << " not compatible with kernel compiled for "
-              << conv2d::kParallelPixels
-              << " pixels in parallel. Please choose a different image, or "
-                 "recompile "
-                 "with a different value of the PARALLEL_PIXELS "
-                 "pre-processor macro."
-              << std::endl;
-    return false;
-  }
-
-  return image_size_ok;
-}
-
 #if TEST_CONV2D_ISOLATED
+
 constexpr std::array<float, 9> identity_coeffs = {
     0.0f, 0.0f, 0.0f,  //
     0.0f, 1.0f, 0.0f,  //
@@ -193,25 +84,13 @@ bool TestTinyFrameOnStencil(sycl::queue q, bool print_debug_info) {
       102, 202, 302, 402, 502, 602, 702, 802,  //
       103, 203, 303, 403, 503, 603, 703, 803};
 
-  vvp_stream_adapters::WriteFrameToPipe<InputImageStreamGrey>(
-      q, rows_small, cols_small, grey_pixels_in);
-
-  // add extra pixels to flush out the FIFO after all image frames
-  // have been added
-  int dummy_pixels = cols_small * conv2d::kWindowSize;
-  vvp_stream_adapters::WriteDummyPixelsToPipe<InputImageStreamGrey>(
-      q, dummy_pixels, (uint16_t)15);
-
-  sycl::event e = q.single_task<ID_Convolution2d>(
-      Convolution2d<InputImageStreamGrey, OutputImageStreamGrey>{
-          (int)rows_small, (int)cols_small, identity_coeffs});
-
   conv2d::PixelType grey_pixels_out[pixels_count];
   bool sidebands_ok;
   int parsed_frames;
-  vvp_stream_adapters::ReadFrameFromPipe<OutputImageStreamGrey>(
-      q, rows_small, cols_small, grey_pixels_out, sidebands_ok, parsed_frames,
-      print_debug_info);
+
+  KernelProcessSingleFrame(q, grey_pixels_in, grey_pixels_out, rows_small,
+                           cols_small, identity_coeffs, sidebands_ok,
+                           parsed_frames);
 
   bool pixels_match = true;
   for (int i = 0; i < pixels_count; i++) {
@@ -223,12 +102,7 @@ bool TestTinyFrameOnStencil(sycl::queue q, bool print_debug_info) {
     pixels_match &= (grey_pixel_expected == grey_pixels_out[i]);
   }
 
-  // Stop the kernel in case testbench wants to run again with different kernel
-  // arguments.
-  StopCSR::write(q, true);
-  e.wait();
-
-  return sidebands_ok & pixels_match;
+  return sidebands_ok & pixels_match & (parsed_frames == 1);
 }
 
 /// @brief Test that the 'bypass' control works correctly.
@@ -252,28 +126,16 @@ bool TestBypass(sycl::queue q, bool print_debug_info) {
       102, 202, 302, 402, 502, 602, 702, 802,  //
       103, 203, 303, 403, 503, 603, 703, 803};
 
-  vvp_stream_adapters::WriteFrameToPipe<InputImageStreamGrey>(
-      q, rows_small, cols_small, grey_pixels_in);
-
-  // add extra pixels to flush out the FIFO after all image frames
-  // have been added
-  int dummy_pixels = cols_small * conv2d::kWindowSize;
-  vvp_stream_adapters::WriteDummyPixelsToPipe<InputImageStreamGrey>(
-      q, dummy_pixels, (uint16_t)15);
-
   // Enable 'bypass' mode by writing to CSR.
   BypassCSR::write(q, true);
-
-  sycl::event e = q.single_task<ID_Convolution2d>(
-      Convolution2d<InputImageStreamGrey, OutputImageStreamGrey>{
-          (int)rows_small, (int)cols_small, identity_coeffs});
 
   conv2d::PixelType grey_pixels_out[pixels_count];
   bool sidebands_ok;
   int parsed_frames;
-  vvp_stream_adapters::ReadFrameFromPipe<OutputImageStreamGrey>(
-      q, rows_small, cols_small, grey_pixels_out, sidebands_ok, parsed_frames,
-      print_debug_info);
+
+  KernelProcessSingleFrame(q, grey_pixels_in, grey_pixels_out, rows_small,
+                           cols_small, identity_coeffs, sidebands_ok,
+                           parsed_frames);
 
   bool pixels_match = true;
   for (int i = 0; i < pixels_count; i++) {
@@ -281,12 +143,7 @@ bool TestBypass(sycl::queue q, bool print_debug_info) {
     pixels_match &= (grey_pixel_expected == grey_pixels_out[i]);
   }
 
-  // Stop the kernel in case testbench wants to run again with different kernel
-  // arguments.
-  StopCSR::write(q, true);
-  e.wait();
-
-  return sidebands_ok & pixels_match;
+  return sidebands_ok & pixels_match & (parsed_frames == 1);
 }
 
 #else
@@ -321,10 +178,7 @@ bool TestGoodFramesSequence(sycl::queue q, size_t num_frames,
             << "\n**********************************\n"
             << std::endl;
 
-  sycl::event e;
-  bool all_passed = true;
-  size_t rows = 0, cols = 0;
-
+  std::vector<bmp_tools::BitmapRGB> input_bitmaps;
   for (size_t itr = 0; itr < num_frames; itr++) {
     // load image
     std::string input_bmp_path =  //
@@ -332,58 +186,19 @@ bool TestGoodFramesSequence(sycl::queue q, size_t num_frames,
 
     std::cout << "INFO: Load image " << input_bmp_path << std::endl;
     unsigned int error;
-    bmp_tools::BitmapRGB in_img = bmp_tools::ReadBmp(input_bmp_path, error);
+    input_bitmaps.push_back(bmp_tools::ReadBmp(input_bmp_path, error));
     if (bmp_tools::BmpError::OK != error) {
       std::cerr << "ERROR: Could not read image from " << input_bmp_path
                 << std::endl;
       return false;
     }
-
-    size_t rows_new = in_img.GetRows();
-    size_t cols_new = in_img.GetRows();
-
-    bool image_ok =
-        UpdateAndCheckImageDimensions(rows, cols, rows_new, cols_new);
-
-    if (!image_ok) {
-      std::cerr << "ERROR: invalid image size " << rows << " x " << cols
-                << std::endl;
-      continue;
-    }
-
-    conv2d::PixelRGB *in_img_vvp = new conv2d::PixelRGB[rows * cols];
-
-    ConvertToVvpRgb(in_img, in_img_vvp, rows * cols);
-
-    vvp_stream_adapters::WriteFrameToPipe<InputImageStream>(q, rows, cols,
-                                                            in_img_vvp);
-
-    // don't need in_img_vvp anymore
-    delete[] in_img_vvp;
   }
 
-  // extra pixels to flush out the FIFO
-  int dummy_pixels = cols * (conv2d::kWindowSize - 1);
-  constexpr auto kDummyVal = conv2d::PixelRGB{100, 100, 100};
-  vvp_stream_adapters::WriteDummyPixelsToPipe<InputImageStream,
-                                              conv2d::PixelRGB>(q, dummy_pixels,
-                                                                kDummyVal);
+  bool all_passed = false;
+  std::vector<bmp_tools::BitmapRGB> output_bitmaps =
+      SystemProcessFrameSequence(q, input_bitmaps, sobel_coeffs, all_passed);
 
-  std::cout << "\n*********************" << std::endl;
-  std::cout << "Launch RGB2Grey kernel" << std::endl;
-  q.single_task<ID_RGB2Grey>(
-      RGB2Grey<InputImageStream, InputImageStreamGrey>{});
-
-  std::cout << "Launch Convolution2d kernel" << std::endl;
-  e = q.single_task<ID_Convolution2d>(
-      Convolution2d<InputImageStreamGrey, OutputImageStreamGrey>{
-          (int)rows, (int)cols, sobel_coeffs});
-
-  std::cout << "Launch Grey2RGB kernel" << std::endl;
-  q.single_task<ID_Grey2RGB>(
-      Grey2RGB<OutputImageStreamGrey, OutputImageStream>{});
-
-  for (size_t itr = 0; itr < num_frames; itr++) {
+  for (size_t itr = 0; itr < output_bitmaps.size(); itr++) {
     std::cout << "\n*********************\n"  //
               << "Reading out frame " << itr  //
               << std::endl;
@@ -391,27 +206,9 @@ bool TestGoodFramesSequence(sycl::queue q, size_t num_frames,
     std::string absolute_output_bmp_path = output_bmp_filename_base + "_" +
                                            std::to_string(itr) +
                                            DEFAULT_EXTENSION;
-
-    conv2d::PixelRGB *out_img_vvp = new conv2d::PixelRGB[rows * cols];
-    InitializeBuffer(out_img_vvp, rows * cols);
-
-    int parsed_frames = 0;
-    bool sidebands_ok = false;
-    if (out_img_vvp) {
-      vvp_stream_adapters::ReadFrameFromPipe<OutputImageStream>(
-          q, rows, cols, out_img_vvp, sidebands_ok, parsed_frames,
-          print_debug_messages);
-    }
-
-    if (1 != parsed_frames) {
-      std::cerr << "ERROR: saw " << parsed_frames
-                << " parsed frames (expected 1)." << std::endl;
-    }
-    bmp_tools::BitmapRGB out_img = ConvertToBmpRgb(out_img_vvp, rows, cols);
-    delete[] out_img_vvp;
-
     unsigned int write_error;
-    bmp_tools::WriteBmp(absolute_output_bmp_path, out_img, write_error);
+    bmp_tools::WriteBmp(absolute_output_bmp_path, output_bitmaps[itr],
+                        write_error);
 
     if (bmp_tools::BmpError::OK == write_error) {
       std::cout << "Wrote convolved image " << absolute_output_bmp_path
@@ -423,12 +220,10 @@ bool TestGoodFramesSequence(sycl::queue q, size_t num_frames,
 
     std::cout << "Compare with " << expected_bmp_path << ". " << std::endl;
     unsigned int comparison_error = 0;
-    bool passed =
-        bmp_tools::CompareFrames(out_img, expected_bmp_path, comparison_error);
+    bool passed = bmp_tools::CompareFrames(output_bitmaps[itr],
+                                           expected_bmp_path, comparison_error);
 
-    all_passed &= passed & sidebands_ok & (1 == parsed_frames);
-    printf("frame %zu %s\n", itr,
-           (passed && sidebands_ok) ? "passed" : "failed");
+    printf("frame %zu %s\n", itr, (passed && all_passed) ? "passed" : "failed");
   }
 
   int detected_version = VersionCSR::read(q);
@@ -439,11 +234,6 @@ bool TestGoodFramesSequence(sycl::queue q, size_t num_frames,
     std::cerr << "ERROR: kernel version did not match!" << std::endl;
     all_passed = false;
   }
-
-  // Stop the kernel in case testbench wants to run again with different kernel
-  // arguments.
-  StopCSR::write(q, true);
-  e.wait();
 
   std::cout << "\nFinished checking a sequence of good frames.\n\n"
             << std::endl;
@@ -487,82 +277,24 @@ bool TestDefectiveFrame(sycl::queue q, std::string input_bmp_filename,
     return false;
   }
 
-  size_t rows_new = in_img.GetRows();
-  size_t cols_new = in_img.GetRows();
-
-  size_t rows = 0;
-  size_t cols = 0;
-  bool image_ok = UpdateAndCheckImageDimensions(rows, cols, rows_new, cols_new);
-
-  if (!image_ok) {
-    std::cerr << "ERROR: invalid image size " << rows << " x " << cols
-              << std::endl;
-  }
-
-  int end_pixel = rows * cols / 2;
-
-  // Enqueue a defective frame that ends after `end_pixel` pixels.
-  conv2d::PixelRGB *in_img_vvp = new conv2d::PixelRGB[rows * cols];
-  conv2d::PixelRGB *out_img_vvp = new conv2d::PixelRGB[rows * cols];
-  ConvertToVvpRgb(in_img, in_img_vvp, rows * cols);
-
-  vvp_stream_adapters::WriteFrameToPipe<InputImageStream>(
-      q, rows, cols, in_img_vvp, end_pixel);
-
-  // Now enqueue a good frame.
-  vvp_stream_adapters::WriteFrameToPipe<InputImageStream>(q, rows, cols,
-                                                          in_img_vvp);
-
-  int dummy_pixels = cols * conv2d::kWindowSize;
-  vvp_stream_adapters::WriteDummyPixelsToPipe<InputImageStream>(
-      q, dummy_pixels, conv2d::PixelRGB{32, 32, 32});
-
-  // Enqueue the kernel. Run it until we have read out the partial frame and
-  // good frame, then stop.
-  sycl::event e;
-
-  std::cout << "\n*********************" << std::endl;
-  std::cout << "Launch RGB2Grey kernel" << std::endl;
-  q.single_task<ID_RGB2Grey>(
-      RGB2Grey<InputImageStream, InputImageStreamGrey>{});
-
-  std::cout << "Launch Convolution2d kernel" << std::endl;
-  e = q.single_task<ID_Convolution2d>(
-      Convolution2d<InputImageStreamGrey, OutputImageStreamGrey>{
-          (int)rows, (int)cols, sobel_coeffs});
-
-  std::cout << "Launch Grey2RGB kernel" << std::endl;
-  q.single_task<ID_Grey2RGB>(
-      Grey2RGB<OutputImageStreamGrey, OutputImageStream>{});
-
-  std::string defect_output_bmp_path =
-      output_bmp_filename_base + "_defect" + DEFAULT_EXTENSION;
-
-  InitializeBuffer(out_img_vvp, rows * cols);
-
-  std::cout << "\n****************************\n"  //
-            << "Read out defective frame, and overwrite with good frame."
-            << std::endl;
-
-  bool sidebands_ok = false;
-  int parsed_frames = 0;
-  if (out_img_vvp) {
-    vvp_stream_adapters::ReadFrameFromPipe<OutputImageStream>(
-        q, rows, cols, out_img_vvp, sidebands_ok, parsed_frames,
-        print_debug_messages);
-  }
   bool passed = true;
 
+  bool sidebands_ok = 0;
+  int num_parsed_frames = 0;
+  bmp_tools::BitmapRGB parsed_frame = SystemProcessFrameAndDefect(
+      q, in_img, sobel_coeffs, num_parsed_frames, sidebands_ok);
+
   // expect the defective frame + the good frame
-  if (2 != parsed_frames) {
-    std::cerr << "ERROR: saw " << parsed_frames
+  if (2 != num_parsed_frames) {
+    std::cerr << "ERROR: saw " << num_parsed_frames
               << " parsed frames (expected 2)." << std::endl;
     passed = false;
   }
 
-  bmp_tools::BitmapRGB out_img = ConvertToBmpRgb(out_img_vvp, rows, cols);
+  std::string defect_output_bmp_path =
+      output_bmp_filename_base + "_defect" + DEFAULT_EXTENSION;
   unsigned int write_error;
-  bmp_tools::WriteBmp(defect_output_bmp_path, out_img, write_error);
+  bmp_tools::WriteBmp(defect_output_bmp_path, parsed_frame, write_error);
 
   if (bmp_tools::BmpError::OK == write_error) {
     std::cout << "Wrote convolved image " << defect_output_bmp_path
@@ -571,19 +303,11 @@ bool TestDefectiveFrame(sycl::queue q, std::string input_bmp_filename,
   // This should succeed since the defective pixels were overwritten by the
   // subsequent good frame.
   unsigned int comparison_error = 0;
-  passed &=
-      bmp_tools::CompareFrames(out_img, expected_bmp_path, comparison_error);
+  passed &= bmp_tools::CompareFrames(parsed_frame, expected_bmp_path,
+                                     comparison_error);
 
   bool all_passed = passed & sidebands_ok;
   printf("frame 'defect' %s\n", (passed && sidebands_ok) ? "passed" : "failed");
-
-  // Stop the kernel in case testbench wants to run again with different kernel
-  // arguments.
-  StopCSR::write(q, true);
-  e.wait();
-
-  delete[] out_img_vvp;
-  delete[] in_img_vvp;
 
   return all_passed;
 }
